server/requirements.txt
appnope==0.1.0
decorator==4.0.2
Django==1.8.4
django-codemirror-widget==0.4.0
django-flatblocks==0.9.1
django-grappelli==2.7.1
django-modeltranslation==0.10
django-multiselectfield==0.1.3
django-wysiwyg-redactor==0.4.9
easy-thumbnails==2.2
funcy==1.5
gnureadline==6.3.3
gunicorn==19.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
modeltranslation-grappelli==3.0.0
path.py==7.6
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.6.1
pytils==0.3
pytz==2015.4
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
